requirements.txt
openai>=1.0.0
feedparser
requests
feedgen
python-dotenv>=1.0.0
pandas
qdrant-client>=1.6.0